Restructure modules into packages
