+++ data_sentinel/modules/__init__.py
@@ -1,0 +1,11 @@
+"""Collection of built-in pipeline modules."""
+
+from . import readers, writers, ops
+from .readers import *  # noqa: F401,F403
+from .writers import *  # noqa: F401,F403
+from .ops import *  # noqa: F401,F403
+
+__all__ = []
+__all__ += readers.__all__
+__all__ += writers.__all__
+__all__ += ops.__all__
