+++ data_sentinel/modules/ops/__init__.py
@@ -1,0 +1,3 @@
+from .dq import DataQualityModule
+
+__all__ = ["DataQualityModule"]
