+++ data_sentinel/modules/ops/dq.py
@@ -1,5 +1,4 @@
-from ..base import BaseModule
-
+from ...base import BaseModule
 
 class DataQualityModule(BaseModule):
     """Simple data quality check that ensures input is not empty."""
@@ -8,5 +7,4 @@ class DataQualityModule(BaseModule):
         allow_empty = self.config.get("allow_empty", False)
         if not allow_empty and not data:
             raise ValueError("Received empty data")
-        # Return data unchanged for this placeholder
         return data
