+++ data_sentinel/modules/readers/__init__.py
@@ -1,0 +1,5 @@
+from .base import BaseReader
+from .realtime import RealTimeReader
+from .s3 import S3Reader
+
+__all__ = ["BaseReader", "RealTimeReader", "S3Reader"]
