+++ data_sentinel/modules/readers/base.py
@@ -1,0 +1,7 @@
+from ...base import BaseModule
+
+class BaseReader(BaseModule):
+    """Base class for all reader modules."""
+
+    def run(self, data=None):
+        raise NotImplementedError
