+++ data_sentinel/modules/readers/s3.py
@@ -1,0 +1,9 @@
+from .base import BaseReader
+
+class S3Reader(BaseReader):
+    """Example S3 reader (placeholder implementation)."""
+
+    def run(self, data=None):
+        bucket = self.config.get("bucket")
+        key = self.config.get("key")
+        return {"source": f"s3://{bucket}/{key}"}
