+++ data_sentinel/modules/writers/__init__.py
@@ -1,0 +1,4 @@
+from .base import BaseWriter
+from .console import ConsoleWriter
+
+__all__ = ["BaseWriter", "ConsoleWriter"]
