+++ data_sentinel/modules/writers/base.py
@@ -1,0 +1,7 @@
+from ...base import BaseModule
+
+class BaseWriter(BaseModule):
+    """Base class for all writer modules."""
+
+    def run(self, data):
+        raise NotImplementedError
